+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+Version **1.1.0**
+
+Date Released: **12/18/2025**
+
+- Support for opamp generation with multiple subcircuits
+- Current and voltage bias source customization
+- Random net customization
+- Improved code readability and organization
+
+
 Version **1.0.0**
 
 Date Released: **11/24/2025**